move snake into its own folder, remove snake from routes
