+++ client/routes.js
@@ -8,7 +8,6 @@ import LandingPage from './components/LandingPage';
 import GameOver from './components/GameOver';
 import Profile from './components/Profile';
 import Leadership from './components/Leadership';
-import SnakeGame from './components/SnakeGame';
 
 /**
  * COMPONENT
@@ -25,20 +24,19 @@ class Routes extends Component {
       <div>
         {isLoggedIn ? (
           <Switch>
-            <Route path='/landing' component={LandingPage} />
-            <Route path='/home' component={Home} />
-            <Route path='/gameover' component={GameOver} />
-            <Route path='/profile' component={Profile} />
-            <Route path='/snakegame' component={SnakeGame} />
-            <Route path='/leadership' component={Leadership} />
-            <Redirect to='/landing' />
+            <Route path="/landing" component={LandingPage} />
+            <Route path="/home" component={Home} />
+            <Route path="/gameover" component={GameOver} />
+            <Route path="/profile" component={Profile} />
+            <Route path="/leadership" component={Leadership} />
+            <Redirect to="/landing" />
           </Switch>
         ) : (
           <Switch>
-            <Route path='/' exact component={Login} />
-            <Route path='/login' component={Login} />
-            <Route path='/signup' component={Signup} />
-            <Redirect to='/login' />
+            <Route path="/" exact component={Login} />
+            <Route path="/login" component={Login} />
+            <Route path="/signup" component={Signup} />
+            <Redirect to="/login" />
           </Switch>
         )}
       </div>
